move Herbel model into redshift module
